refactor(tui): split stack, box, shapes into internal modules
Split stack.ts (871→41), box.ts (641→372), shapes.ts (500→196).
Internal modules at src/render/internal/ for types, helpers,
composition, guards, and registry.

🤘🏻 In-collaboration-with: [Claude Code](https://claude.com/claude-code)
